Extract RefreshRemoteTask
