+++ k9mail/src/main/java/com/fsck/k9/controller/MessagingController.java
@@ -61,6 +61,7 @@ import com.fsck.k9.controller.MessagingControllerCommands.PendingSetFlag;
 import com.fsck.k9.controller.ProgressBodyFactory.ProgressListener;
 import com.fsck.k9.controller.imap.ImapMessageStore;
 import com.fsck.k9.controller.tasks.ListFoldersTask;
+import com.fsck.k9.controller.tasks.RefreshRemoteTask;
 import com.fsck.k9.helper.Contacts;
 import com.fsck.k9.mail.Address;
 import com.fsck.k9.mail.AuthenticationFailedException;
@@ -365,88 +366,16 @@ public class MessagingController implements IMessageController {
     }
 
     @Override
-    public void listFoldersSynchronous(final Account account, final boolean refreshRemote, final MessagingListener listener) {
+    public void listFoldersSynchronous(final Account account, final boolean refreshRemote,
+            final MessagingListener listener) {
         new ListFoldersTask(this,
                 context, account, refreshRemote, listener, listeners).run();
     }
 
     @Override
     public void doRefreshRemote(final Account account, final MessagingListener listener) {
-        put("doRefreshRemote", listener, new Runnable() {
-            @Override
-            public void run() {
-                refreshRemoteSynchronous(account, listener);
-            }
-        });
-    }
-
-    @VisibleForTesting
-    void refreshRemoteSynchronous(final Account account, final MessagingListener listener) {
-        List<LocalFolder> localFolders = null;
-        try {
-            Store store = account.getRemoteStore();
-
-            List<? extends Folder> remoteFolders = store.getPersonalNamespaces(false);
-
-            LocalStore localStore = account.getLocalStore();
-            Set<String> remoteFolderNames = new HashSet<>();
-            List<LocalFolder> foldersToCreate = new LinkedList<>();
-
-            localFolders = localStore.getPersonalNamespaces(false);
-            Set<String> localFolderNames = new HashSet<>();
-            for (Folder localFolder : localFolders) {
-                localFolderNames.add(localFolder.getName());
-            }
-
-            for (Folder remoteFolder : remoteFolders) {
-                if (!localFolderNames.contains(remoteFolder.getName())) {
-                    LocalFolder localFolder = localStore.getFolder(remoteFolder.getName());
-                    foldersToCreate.add(localFolder);
-                }
-                remoteFolderNames.add(remoteFolder.getName());
-            }
-            localStore.createFolders(foldersToCreate, account.getDisplayCount());
-
-            localFolders = localStore.getPersonalNamespaces(false);
-
-            /*
-             * Clear out any folders that are no longer on the remote store.
-             */
-            for (Folder localFolder : localFolders) {
-                String localFolderName = localFolder.getName();
-
-                // FIXME: This is a hack used to clean up when we accidentally created the
-                //        special placeholder folder "-NONE-".
-                if (K9.FOLDER_NONE.equals(localFolderName)) {
-                    localFolder.delete(false);
-                }
-
-                if (!account.isSpecialFolder(localFolderName) &&
-                        !remoteFolderNames.contains(localFolderName)) {
-                    localFolder.delete(false);
-                }
-            }
-
-            localFolders = localStore.getPersonalNamespaces(false);
-
-            for (MessagingListener l : getListeners(listener)) {
-                l.listFolders(account, localFolders);
-            }
-            for (MessagingListener l : getListeners(listener)) {
-                l.listFoldersFinished(account);
-            }
-        } catch (Exception e) {
-            for (MessagingListener l : getListeners(listener)) {
-                l.listFoldersFailed(account, "");
-            }
-            Timber.e(e);
-        } finally {
-            if (localFolders != null) {
-                for (Folder localFolder : localFolders) {
-                    closeFolder(localFolder);
-                }
-            }
-        }
+        put("doRefreshRemote", listener,
+                new RefreshRemoteTask(account, listener, listeners));
     }
 
     /**
