+++ k9mail/src/test/java/com/fsck/k9/controller/MessagingControllerTest.java
@@ -200,106 +200,6 @@ public class MessagingControllerTest {
     }
 
     @Test
-    public void refreshRemoteSynchronous_shouldCreateFoldersFromRemote() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        LocalFolder newLocalFolder = mock(LocalFolder.class);
-
-        List<Folder> folders = Collections.singletonList(remoteFolder);
-        when(remoteStore.getPersonalNamespaces(false)).thenAnswer(createAnswer(folders));
-        when(remoteFolder.getName()).thenReturn("NewFolder");
-        when(localStore.getFolder("NewFolder")).thenReturn(newLocalFolder);
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(localStore).createFolders(eq(Collections.singletonList(newLocalFolder)), anyInt());
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldDeleteFoldersNotOnRemote() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        LocalFolder oldLocalFolder = mock(LocalFolder.class);
-        when(oldLocalFolder.getName()).thenReturn("OldLocalFolder");
-        when(localStore.getPersonalNamespaces(false))
-                .thenReturn(Collections.singletonList(oldLocalFolder));
-        List<Folder> folders = Collections.emptyList();
-        when(remoteStore.getPersonalNamespaces(false)).thenAnswer(createAnswer(folders));
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(oldLocalFolder).delete(false);
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldNotDeleteFoldersOnRemote() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        when(localStore.getPersonalNamespaces(false))
-                .thenReturn(Collections.singletonList(localFolder));
-        List<Folder> folders = Collections.singletonList(remoteFolder);
-        when(remoteStore.getPersonalNamespaces(false)).thenAnswer(createAnswer(folders));
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(localFolder, never()).delete(false);
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldNotDeleteSpecialFoldersNotOnRemote() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        LocalFolder missingSpecialFolder = mock(LocalFolder.class);
-        when(account.isSpecialFolder("Outbox")).thenReturn(true);
-        when(missingSpecialFolder.getName()).thenReturn("Outbox");
-        when(localStore.getPersonalNamespaces(false))
-                .thenReturn(Collections.singletonList(missingSpecialFolder));
-        List<Folder> folders = Collections.emptyList();
-        when(remoteStore.getPersonalNamespaces(false)).thenAnswer(createAnswer(folders));
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(missingSpecialFolder, never()).delete(false);
-    }
-
-    public static <T> Answer<T> createAnswer(final T value) {
-        return new Answer<T>() {
-            @Override
-            public T answer(InvocationOnMock invocation) throws Throwable {
-                return value;
-            }
-        };
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldProvideFolderList() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        List<LocalFolder> folders = Collections.singletonList(localFolder);
-        when(localStore.getPersonalNamespaces(false)).thenReturn(folders);
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(listener).listFolders(account, folders);
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldNotifyFinishedAfterSuccess() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        List<LocalFolder> folders = Collections.singletonList(localFolder);
-        when(localStore.getPersonalNamespaces(false)).thenReturn(folders);
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(listener).listFoldersFinished(account);
-    }
-
-    @Test
-    public void refreshRemoteSynchronous_shouldNotNotifyFinishedAfterFailure() throws MessagingException {
-        configureRemoteStoreWithFolder();
-        when(localStore.getPersonalNamespaces(false)).thenThrow(new MessagingException("Test"));
-
-        controller.refreshRemoteSynchronous(account, listener);
-
-        verify(listener, never()).listFoldersFinished(account);
-    }
-
-    @Test
     public void searchLocalMessagesSynchronous_shouldCallSearchForMessagesOnLocalStore()
             throws Exception {
         setAccountsInPreferences(Collections.singletonMap("1", account));
